requirements.txt
altgraph==0.17
fbs==0.9.0
ffprobe-python==1.0.3
future==0.18.2
macholib==1.14
numpy==1.19.5
opencv-contrib-python==4.5.1.48
pefile==2019.4.18
PyInstaller==3.4
PySide2==5.15.2
shiboken2==5.15.2
